+++ src/server/server_create.c
@@ -8,7 +8,7 @@
 #include <stddef.h>
 #include <stdlib.h>
 
-knServer *knServer_create(size_t port)
+knServer *knServer_create(uint16_t port)
 {
     knServer *server = calloc(1, sizeof(knServer));
